Move Ui, Storage and Parser into package duke.main
