+++ data/duke.txt
@@ -1,1 +1,1 @@
-D | 0 | rewer | rwer
+D | 0 | dedade | erterte
