+++ src/main/java/duke/command/FindCommand.java
@@ -1,9 +1,9 @@
 package duke.command;
 
-import duke.storage.Storage;
+import duke.main.Storage;
 import duke.task.Task;
 import duke.task.TaskList;
-import duke.ui.Ui;
+import duke.main.Ui;
 import java.util.ArrayList;
 
 /**
