+++ src/main/java/duke/main/Storage.java
@@ -1,4 +1,4 @@
-package duke.storage;
+package duke.main;
 
 import duke.task.Task;
 import duke.exception.DukeException;
