+++ src/main/java/duke/main/Ui.java
@@ -1,4 +1,4 @@
-package duke.ui;
+package duke.main;
 
 import duke.task.Task;
 import java.util.ArrayList;
